+++ package.json
@@ -16,6 +16,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-icons": "^5.5.0",
+    "react-spinners": "^0.17.0",
     "react-toastify": "^11.0.5",
     "zustand": "^5.0.8"
   },
