+++ src/app/interview-setup/constants/interviewRules.constants.ts
